requirements.txt
torch>=2.0.0
diffusers>=0.26.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
Pillow>=9.0.0
opencv-python>=4.5.0
torchvision>=0.15.0
protobuf>=3.20.0
sentencepiece>=0.1.97
# Development tools
pre-commit>=3.5.0
black>=24.3.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
